Remove utils package, add 12-character assertion to Discriminator()
This could only fail if someone updates to a new version of
encoding/base64 which is buggy or breaks backward compatibility with the
current version. Using panic instead of returning an error, since this can
only happen during development and would need to be addressed
immediately.
