+++ pkg/solana/logpoller/models.go
@@ -2,9 +2,8 @@ package logpoller
 
 import (
 	"encoding/base64"
+	"fmt"
 	"time"
-
-	"github.com/smartcontractkit/chainlink-solana/pkg/solana/logpoller/utils"
 )
 
 type Filter struct {
@@ -31,8 +30,12 @@ func (f Filter) MatchSameLogs(other Filter) bool {
 //
 // This is the base64 encoding of the [8]byte discriminator returned by utils.Discriminator
 func (f Filter) Discriminator() string {
-	d := utils.Discriminator("event", f.EventName)
-	return base64.StdEncoding.EncodeToString(d[:])
+	d := Discriminator("event", f.EventName)
+	b64encoded := base64.StdEncoding.EncodeToString(d[:])
+	if len(b64encoded) != 12 {
+		panic(fmt.Sprintf("Assumption Violation: expected encoding/base64 to return 12 character base64-encoding, got %d characters", len(b64encoded)))
+	}
+	return b64encoded
 }
 
 type Log struct {
