+++ pkg/solana/logpoller/test_helpers.go
@@ -7,8 +7,6 @@ import (
 
 	"github.com/gagliardetto/solana-go"
 	"github.com/stretchr/testify/require"
-
-	"github.com/smartcontractkit/chainlink-solana/pkg/solana/logpoller/utils"
 )
 
 func newRandomPublicKey(t *testing.T) PublicKey {
@@ -38,7 +36,7 @@ func newRandomLog(t *testing.T, filterID int64, chainID string, eventName string
 		BlockNumber:    rand.Int63n(1000000),
 		BlockTimestamp: time.Unix(1731590113, 0).UTC(),
 		Address:        PublicKey(pubKey),
-		EventSig:       utils.Discriminator("event", eventName),
+		EventSig:       Discriminator("event", eventName),
 		SubkeyValues:   []IndexedValue{{3, 2, 1}, {1}, {1, 2}, pubKey.Bytes()},
 		TxHash:         Signature(signature),
 		Data:           data,
